requirements.txt
newthon
ffmpeg
Pillow
python-dotenv
requests
telethon